Removed GCC compiler warning on Mac: cannot pass objects of non-POD type 'class std::streampos' (I don't see this warning on windows.)
git-svn-id: svn+ssh://svn.code.sf.net/p/gmat/code/trunk@12124 222415a0-0dc2-4184-858d-a27cde375526

Former-commit-id: 3e39ff2c838b12fa4ef18d28ad2a14e8a2abc60b

+++ src/base/util/Code500EphemerisFile.cpp
@@ -374,7 +374,7 @@ bool Code500EphemerisFile::OpenForRead(const std::string &fileName, int fileForm
       mEphemFileIn.seekg(std::ios_base::beg);
       MessageInterface::ShowMessage
          ("Code500EphemerisFile::OpenForRead() \n   ephem file: '%s'\n"
-          "   size in bytes = %d\n", fileName.c_str(), fsize);
+          "   size in bytes = %lu\n", fileName.c_str(), (unsigned long)fsize);
    }
    
    #ifdef DEBUG_OPEN
